Share code online With Oscript Editor. (#10)
* Share code online With Oscript Editor

Added a sharing code button online,
Users can share the code from oscript editor to directly show it's working to anyone. The code sharing is powered with free api http://myjson.com which will took the whole agent object and create a router url for it.

* Showing copy to clipboard

Added `navigator.clipboard.writeText` which either requires a secure protocol (Like https or localhost) To work otherwise it will show an error (Alert users if copied). Using this method is easy to manage than using document.execCommand('copy')  which require to create a new hidden HTML element and then destroy it.

* added notification for user and copying to clipboard

* Fixed duplicating same text aa

Fixed When creating duplicate aa's label even if text is same,
Now if a label is shared then a new -id filed with append in label when sharing it again with different text may result in `label -id -id` and so on.

* adding serial number

added serial number for same label name

* adding new label name

Solved error that was preventing from adding new label name.

* Solved issues pointed by Anton

Solved possible issues those are reviewed by Anton (Obyte Founder) and which helps to improve the code now and in the future.

* removed electron stuff

Removed electron stuff from package.json